requirements.txt
asgiref==3.2.10
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
Django==3.1.5
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.9
imageio==1.5
importlib-metadata==1.6.0
itchat==1.3.10
itypes==1.2.0
Jinja2==2.11.2
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
optionaldict==0.1.2
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.10.1
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
tinydb==4.3.0
uritemplate==3.0.1
urllib3==1.25.9
wechatpy==1.8.14
xmltodict==0.12.0
zipp==3.1.0

xlwt~=1.3.0